reorg: Migrate the rest of `unistd.h` to the new module
Link: https://pubs.opengroup.org/onlinepubs/007904975/basedefs/unistd.h.html
