+++ src/new/hermit_abi/hermit_mod.rs
@@ -1,0 +1,3 @@
+//! Hermit kernel libc.
+//!
+//! * Headers: <https://github.com/hermit-os/hermit-rs/tree/main/hermit-abi>
