+++ src/new/hurd/mod.rs
@@ -1,2 +1,4 @@
 //! GNU Hurd libc.
+//!
+//! Hurd interfaces come from the glibc module.
 // FIXME(hurd): link to headers needed.
