+++ src/new/sgx/unistd.rs
@@ -1,6 +1,73 @@
 //! Header: `unistd.h`
 
 pub use crate::new::common::posix::unistd::{
+    _exit,
+    access,
+    alarm,
+    chdir,
+    chown,
+    close,
+    confstr,
+    dup,
+    dup2,
+    execl,
+    execle,
+    execlp,
+    execv,
+    execve,
+    execvp,
+    fchdir,
+    fchown,
+    fork,
+    fpathconf,
+    fsync,
+    ftruncate,
+    getcwd,
+    getegid,
+    geteuid,
+    getgid,
+    getgroups,
+    gethostname,
+    getlogin,
+    getopt,
+    getpgid,
+    getpgrp,
+    getpid,
+    getppid,
+    getsid,
+    getuid,
+    isatty,
+    lchown,
+    link,
+    lockf,
+    lseek,
+    nice,
+    pathconf,
+    pause,
+    pipe,
+    pread,
+    pwrite,
+    read,
+    readlink,
+    rmdir,
+    setegid,
+    seteuid,
+    setgid,
+    setpgid,
+    setregid,
+    setreuid,
+    setsid,
+    setuid,
+    sleep,
+    symlink,
+    sysconf,
+    tcgetpgrp,
+    tcsetpgrp,
+    truncate,
+    ttyname,
+    ttyname_r,
+    unlink,
+    write,
     STDERR_FILENO,
     STDIN_FILENO,
     STDOUT_FILENO,
