+++ src/new/vxworks/unistd.rs
@@ -1,6 +1,51 @@
 //! Header: `unistd.h`
 
 pub use crate::new::common::posix::unistd::{
+    _exit,
+    access,
+    alarm,
+    chdir,
+    chown,
+    close,
+    dup,
+    dup2,
+    fchdir,
+    fchown,
+    fpathconf,
+    fsync,
+    ftruncate,
+    getcwd,
+    getegid,
+    geteuid,
+    getgid,
+    getgroups,
+    gethostname,
+    getlogin,
+    getopt,
+    getpid,
+    getppid,
+    getuid,
+    isatty,
+    link,
+    lseek,
+    pathconf,
+    pause,
+    pipe,
+    read,
+    readlink,
+    rmdir,
+    setegid,
+    seteuid,
+    setgid,
+    setpgid,
+    setuid,
+    sleep,
+    symlink,
+    sysconf,
+    truncate,
+    ttyname,
+    unlink,
+    write,
     STDERR_FILENO,
     STDIN_FILENO,
     STDOUT_FILENO,
